rename calibration_set module as calibration to match rust side, implement rest of CalibrationSet bindings
